+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- ![SOURCE] Add sources from Directorates General of the European Commission [#1137](https://github.com/nbbrd/sdmx-dl/issues/1137)
+
 ### Fixed
 
 - ![PROVIDER] Fix field overflow in MsalAuthenticator [#1138](https://github.com/nbbrd/sdmx-dl/issues/1138)